+++ GameNinja/Sprite.h
@@ -11,7 +11,7 @@ class CSprite
 protected:
 	int id; //sprite ID
 
-	float left, top, right, bottom; //position of sprite in texture
+	int left, top, right, bottom; //position of sprite in texture
 
 	LPDIRECT3DTEXTURE9 texture;
 
@@ -19,7 +19,7 @@ protected:
 
 public:
 	CSprite() {};
-	CSprite(int id, float left, float top, float right, float bottom, LPDIRECT3DTEXTURE9 tex);
+	CSprite(int id, int left, int top, int right, int bottom, LPDIRECT3DTEXTURE9 tex);
 	void Draw(float x = 0, float y = 0, bool _isFlipHoritzontal = false);
 	void FlipHorizontal(bool flip);
 	//void FlipVertical(bool flip);
@@ -27,8 +27,8 @@ public:
 	//bool isFlipVertical();
 	bool isFlipHorizontal();
 
-	float getWidth();
-	float getHeight();
+	int getWidth();
+	int getHeight();
 };
 
 typedef CSprite * LPSPRITE;
